Move models to lib and split into Schema and Repositories
This is to prepare for Phoenix 1.3 structure where models will go away.
This commit also moves all query code into context modules. This is a
big change in terms of project structure but logic hasn't changed at all
