+++ test/controllers/repository_controller_test.exs
@@ -2,7 +2,8 @@ defmodule Pullhub.RepositoryControllerTest do
   use Pullhub.ConnCase
 
   alias Pullhub.User
-  alias Pullhub.Repository
+  alias Pullhub.Repo
+  alias Pullhub.Repositories
   @valid_attrs %{name: "some content", remote_id: 42}
   @invalid_attrs %{}
 
@@ -22,7 +23,7 @@ defmodule Pullhub.RepositoryControllerTest do
 
   test "lists all entries on index", %{user: user} do
     new_repo = %{remote_id: 11, name: "awesome repo", owner: "1", user_id: user.id}
-              |> Repository.find_or_create
+              |> Repositories.find_or_create
 
     conn = session_conn(user)
     conn = get conn, repository_path(conn, :index)
