+++ web/channels/repository_channel.ex
@@ -21,7 +21,9 @@ defmodule Pullhub.RepositoryChannel do
   def handle_in("update_repositories", %{}, socket) do
     %{id: user_id} = socket.assigns[:user]
 
-    Pullhub.RepoFetcher.fetch(%{user: Pullhub.Repo.get!(Pullhub.User, user_id)})
+    %{user: Pullhub.Accounts.find_by_id!(user_id)}
+    |> Pullhub.RepoFetcher.fetch
+
     {:noreply, socket}
   end
 end
