+++ web/channels/user_socket.ex
@@ -11,7 +11,7 @@ defmodule Pullhub.UserSocket do
   def connect(%{"token" => token}, socket) do
     case Phoenix.Token.verify(socket, "user socket", token, max_age: 1209600) do
       {:ok, user_id} ->
-        {:ok, assign(socket, :user, Pullhub.Repo.get!(Pullhub.User, user_id))}
+        {:ok, assign(socket, :user, Pullhub.Accounts.find_by_id!(user_id))}
       {:error, reason} ->
         :error
     end
