+++ web/controllers/pull_request_controller.ex
@@ -1,9 +1,7 @@
 defmodule Pullhub.PullRequestController do
-  require Logger
   use Pullhub.Web, :controller
 
-  alias Pullhub.Repository
-  alias Pullhub.PullRequest
+  alias Pullhub.Repositories
 
   def index(conn, _params) do
     conn
@@ -17,24 +15,6 @@ defmodule Pullhub.PullRequestController do
 
   def user_repos(conn) do
     user_id(conn)
-    |> Repository.user_repositories
-    |> Repository.enabled_repositories
-    |> preload_pull_requests
-    |> preload_users
-    |> Repo.all
-  end
-
-  def preload_pull_requests(query) do
-    from r in query,
-        preload: [
-          pull_requests: ^Pullhub.PullRequest.open_pull_requests,
-        ]
-  end
-
-  def preload_users(query) do
-    from r in query,
-        preload: [
-          user: :repositories
-        ]
+    |> Repositories.user_repos
   end
 end
